Select Element Height Fix
Currently, there is a style in materialize.css for select elements that limits
the height of select elements to 3rem. This results in the height of select
elements to be restricted down to a size that is really small. This CSS fix
allows the height of the product catalog select elements to fit their contents
more appropriately.